+++ Gruntfile.js
@@ -17,8 +17,32 @@ var SNAP_FILES = [
 
 module.exports = function(grunt) {
 
-    var pkg = grunt.file.readJSON("package.json");
+    var pkg = grunt.file.readJSON("package.json"),
+        core = [
+            "./src/mina.js",
+            "./src/svg.js",
+            "./src/element.js",
+            "./src/animation.js",
+            "./src/matrix.js",
+            "./src/attr.js",
+            "./src/class.js",
+            "./src/attradd.js",
+            "./src/paper.js",
+            "./src/path.js",
+            "./src/set.js",
+            "./src/equal.js",
+            "./src/mouse.js",
+            "./src/filter.js",
+            "./src/align.js",
+            "./src/colors.js"
+        ],
+        src = [
+            "./node_modules/eve/eve.js",
+            "./src/amd-banner.js",
+            "./src/amd-footer.js"
+        ];
 
+    src.splice(2, 0, core);
     // Project configuration.
     grunt.initConfig({
         // Metadata.
@@ -58,12 +82,33 @@ module.exports = function(grunt) {
                 src: [
                     "./src/commonjs-banner.js"
                   ].concat(SNAP_FILES)
-                   .concat(["./src/commonjs-footer.js"])
+                   .concat(["./src/commonjs-footer.js"]),
+                dest: "dist/snap.svg.js",
+                src: src
             }
         },
         exec: {
             dr: {
               command: "node node_modules/dr.js/dr dr.json"
+            },
+            test: {
+              command: "cd test; phantomjs test.js"
+            },
+            eslint: {
+                command: "./node_modules/eslint/bin/eslint.js " + core.join(" ")
+            },
+        },
+        prettify: {
+            options: {
+                indent: 4,
+                indent_char: " ",
+                wrap_line_length: 80,
+                brace_style: "expand",
+                unformatted: ["code", "pre", "script"]
+            },
+            one: {
+                src: "./doc/reference.html",
+                dest: "./doc/reference.html"
             }
         }
     });
@@ -71,6 +116,9 @@ module.exports = function(grunt) {
     grunt.loadNpmTasks("grunt-contrib-concat");
     grunt.loadNpmTasks("grunt-contrib-uglify");
     grunt.loadNpmTasks("grunt-exec");
+    grunt.loadNpmTasks("grunt-prettify");
 
-    grunt.registerTask("default", ["concat", "uglify", "exec"]);
+    grunt.registerTask("default", ["exec:eslint", "concat", "uglify", "exec:dr", "prettify"]);
+    grunt.registerTask("lint", ["exec:eslint"]);
+    grunt.registerTask("test", ["exec:test"]);
 };