+++ src/filter.js
@@ -127,22 +127,23 @@ Snap.plugin(function (Snap, Element, Paper, glob) {
      - opacity (number) #optional `0..1` opacity of the shadow
      = (string) filter representation
      > Usage
-     | var f = paper.filter(Snap.filter.shadow(0, 2, 3)),
+     | var f = paper.filter(Snap.filter.shadow(0, 2, .3)),
      |     c = paper.circle(10, 10, 10).attr({
      |         filter: f
      |     });
     \*/
     Snap.filter.shadow = function (dx, dy, blur, color, opacity) {
-        if (typeof blur == "string") {
-            color = blur;
-            opacity = color;
-            blur = 4;
+        if (opacity == null) {
+            if (color == null) {
+                opacity = blur;
+                blur = 4;
+                color = "#000";
+            } else {
+                opacity = color;
+                color = blur;
+                blur = 4;
+            }
         }
-        if (typeof color != "string") {
-            opacity = color;
-            color = "#000";
-        }
-        color = color || "#000";
         if (blur == null) {
             blur = 4;
         }