+++ src/set.js
@@ -145,6 +145,14 @@ Snap.plugin(function (Snap, Element, Paper, glob) {
             }
         });
     };
+    /*\
+     * Set.remove
+     [ method ]
+     **
+     * Removes all children of the set.
+     *
+     = (object) Set object
+    \*/
     setproto.remove = function () {
         while (this.length) {
             this.pop().remove();
@@ -183,6 +191,13 @@ Snap.plugin(function (Snap, Element, Paper, glob) {
         }
         return this;
     };
+    /*\
+     * Set.attr
+     [ method ]
+     **
+     * Equivalent of @Element.attr.
+     = (object) Set object
+    \*/
     setproto.attr = function (value) {
         var unbound = {};
         for (var k in value) {
@@ -261,6 +276,15 @@ Snap.plugin(function (Snap, Element, Paper, glob) {
         }
         return false;
     };
+    /*\
+     * Set.insertAfter
+     [ method ]
+     **
+     * Inserts set elements after given element.
+     **
+     - element (object) set will be inserted after this element
+     = (object) Set object
+    \*/
     setproto.insertAfter = function (el) {
         var i = this.items.length;
         while (i--) {
@@ -268,6 +292,13 @@ Snap.plugin(function (Snap, Element, Paper, glob) {
         }
         return this;
     };
+    /*\
+     * Set.getBBox
+     [ method ]
+     **
+     * Union of all bboxes of the set. See @Element.getBBox.
+     = (object) bounding box descriptor. See @Element.getBBox.
+    \*/
     setproto.getBBox = function () {
         var x = [],
             y = [],
@@ -295,6 +326,14 @@ Snap.plugin(function (Snap, Element, Paper, glob) {
             cy: y + (y2 - y) / 2
         };
     };
+    /*\
+     * Set.insertAfter
+     [ method ]
+     **
+     * Creates a clone of the set.
+     **
+     = (object) New Set object
+    \*/
     setproto.clone = function (s) {
         s = new Set;
         for (var i = 0, ii = this.items.length; i < ii; i++) {
@@ -307,7 +346,24 @@ Snap.plugin(function (Snap, Element, Paper, glob) {
     };
     setproto.type = "set";
     // export
+    /*\
+     * Snap.Set
+     [ property ]
+     **
+     * Set constructor.
+    \*/
     Snap.Set = Set;
+    /*\
+     * Snap.set
+     [ method ]
+     **
+     * Creates a set and fills it with list of arguments.
+     **
+     = (object) New Set object
+     | var r = paper.rect(0, 0, 10, 10),
+     |     s1 = Snap.set(), // empty set
+     |     s2 = Snap.set(r, paper.circle(100, 100, 20)); // prefilled set
+    \*/
     Snap.set = function () {
         var set = new Set;
         if (arguments.length) {